Update vl6180x_sensor.cpp
fix als_count should be float

+++ esphome/components/vl6180x/vl6180x_sensor.cpp
@@ -300,11 +300,11 @@ float VL6180XSensor::read_als(uint8_t gain) {
   writing_register(VL6180XConstants::VL6180X_REG_SYSALS_START, 0x1);
 
   // Read lux!
-  uint16_t als_count = reading_register16(VL6180XConstants::VL6180X_REG_RESULT_ALS_VAL);
+  float als_count = reading_register16(VL6180XConstants::VL6180X_REG_RESULT_ALS_VAL);
   writing_register(VL6180XConstants::VL6180X_REG_SYSTEM_INTERRUPT_CLEAR, 0x07);
 
   // Log the ALS value
-  ESP_LOGD(TAG, "ALS Count: %hu", als_count);
+  ESP_LOGD(TAG, "ALS Count: %f", als_count);
   //std::string rawhex = format_hex_pretty((uint8_t *) x.c_str(), x.size()).c_str();
   ESP_LOGD(TAG, "ALS Gain: 0x%02X", gain);
